+++ cognite/extractorutils/configtools.py
@@ -85,6 +85,7 @@ Get a state store object as configured:
 However, all of these things will be automatically done for you if you are using the base Extractor class.
 """
 import argparse
+import base64
 import json
 import logging
 import os
@@ -96,16 +97,21 @@ from datetime import timedelta
 from enum import Enum
 from hashlib import sha256
 from logging.handlers import TimedRotatingFileHandler
+from pathlib import Path
 from threading import Event
 from time import sleep
 from typing import Any, Callable, Dict, Generic, Iterable, List, Optional, T, TextIO, Tuple, Type, TypeVar, Union
 from urllib.parse import urljoin
 
+import cryptography.hazmat.primitives.serialization as serialization
+import cryptography.hazmat.primitives.serialization.pkcs12 as pkcs12
 import dacite
 import yaml
 from cognite.client import ClientConfig, CogniteClient
-from cognite.client.credentials import APIKey, OAuthClientCredentials
+from cognite.client.credentials import APIKey, OAuthClientCertificate, OAuthClientCredentials
 from cognite.client.data_classes import Asset, DataSet, ExtractionPipeline
+from cryptography.hazmat.primitives import hashes
+from cryptography.x509 import load_der_x509_certificate, load_pem_x509_certificate
 from prometheus_client import start_http_server
 from prometheus_client.core import REGISTRY
 from yaml.scanner import ScannerError
@@ -243,6 +249,33 @@ class TimeIntervalConfig(yaml.YAMLObject):
         return self._expression
 
 
+def _load_certificate_data(cert_path: str, password: Optional[str]) -> Tuple[str, str]:
+    path = Path(cert_path)
+    cert_data = Path(path).read_bytes()
+
+    if path.suffix == ".pem":
+        cert = load_pem_x509_certificate(cert_data)
+        private_key = serialization.load_pem_private_key(cert_data, password=password.encode() if password else None)
+        private_key_str = private_key.private_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PrivateFormat.TraditionalOpenSSL,
+            encryption_algorithm=serialization.NoEncryption(),
+        )
+        return (base64.b16encode(cert.fingerprint(hashes.SHA1())), private_key_str)
+    elif path.suffix == ".pfx":
+        (private_key, cert, _) = pkcs12.load_key_and_certificates(
+            cert_data, password=password.encode() if password else None
+        )
+        private_key_str = private_key.private_bytes(
+            encoding=serialization.Encoding.PEM,
+            format=serialization.PrivateFormat.TraditionalOpenSSL,
+            encryption_algorithm=serialization.NoEncryption(),
+        )
+        return (base64.b16encode(cert.fingerprint(hashes.SHA1())), private_key_str)
+    else:
+        raise InvalidConfigError(f"Unknown certificate format '{path.suffix}'. Allowed formats are 'pem' and 'pfx'")
+
+
 @dataclass
 class CogniteConfig:
     """
@@ -270,20 +303,42 @@ class CogniteConfig:
         if self.api_key:
             credential_provider = APIKey(self.api_key)
         elif self.idp_authentication:
-            kwargs = {}
-            if self.idp_authentication.token_url:
-                kwargs["token_url"] = self.idp_authentication.token_url
-            elif self.idp_authentication.tenant:
-                base_url = urljoin(self.idp_authentication.authority, self.idp_authentication.tenant)
-                kwargs["token_url"] = f"{base_url}/oauth2/v2.0/token"
-            kwargs["client_id"] = self.idp_authentication.client_id
-            kwargs["client_secret"] = self.idp_authentication.secret
-            kwargs["scopes"] = self.idp_authentication.scopes
-            if token_custom_args is None:
-                token_custom_args = {}
-            if self.idp_authentication.resource:
-                token_custom_args["resource"] = self.idp_authentication.resource
-            credential_provider = OAuthClientCredentials(**kwargs, **token_custom_args)
+            if self.idp_authentication.certificate:
+                if self.idp_authentication.certificate.authority_url:
+                    authority_url = self.idp_authentication.certificate.authority_url
+                elif self.idp_authentication.tenant:
+                    authority_url = urljoin(self.idp_authentication.authority, self.idp_authentication.tenant)
+                else:
+                    raise InvalidConfigError(
+                        "Either authority-url or tenant is required for certificate authentication"
+                    )
+                (thumprint, key) = _load_certificate_data(
+                    self.idp_authentication.certificate.path, self.idp_authentication.certificate.password
+                )
+                credential_provider = OAuthClientCertificate(
+                    authority_url=authority_url,
+                    client_id=self.idp_authentication.client_id,
+                    cert_thumbprint=thumprint,
+                    certificate=key,
+                    scopes=self.idp_authentication.scopes,
+                )
+            elif self.idp_authentication.secret:
+                kwargs = {}
+                if self.idp_authentication.token_url:
+                    kwargs["token_url"] = self.idp_authentication.token_url
+                elif self.idp_authentication.tenant:
+                    base_url = urljoin(self.idp_authentication.authority, self.idp_authentication.tenant)
+                    kwargs["token_url"] = f"{base_url}/oauth2/v2.0/token"
+                kwargs["client_id"] = self.idp_authentication.client_id
+                kwargs["client_secret"] = self.idp_authentication.secret
+                kwargs["scopes"] = self.idp_authentication.scopes
+                if token_custom_args is None:
+                    token_custom_args = {}
+                if self.idp_authentication.resource:
+                    token_custom_args["resource"] = self.idp_authentication.resource
+                credential_provider = OAuthClientCredentials(**kwargs, **token_custom_args)
+            else:
+                raise InvalidConfigError("No client certificate or secret provided")
         else:
             raise InvalidConfigError("No CDF credentials")